// agent-indentity-proxy/src/client/types.gen.ts
export type SapScaiGrantsAuthorizationServiceAuthorizationDetails = {
    ID?: string;
    createdAt?: string | null;
    createdBy?: string | null;
    modifiedAt?: string | null;
    modifiedBy?: string | null;
    transport?: string | null;
    tools?: SapScaiGrantsAuthorizationServiceCdsMap | null;
    server?: string | null;
    roots?: Array<string | null>;
    permissions_read?: boolean | null;
    permissions_write?: boolean | null;
    permissions_execute?: boolean | null;
    permissions_delete?: boolean | null;
    permissions_list?: boolean | null;
    permissions_create?: boolean | null;
    databases?: Array<string | null>;
    schemas?: Array<string | null>;
    tables?: Array<string | null>;
    urls?: Array<string | null>;
    protocols?: Array<string | null>;
    consent?: SapScaiGrantsAuthorizationServiceConsents | null;
    consent_ID?: string | null;
    consent_grant_id?: string | null;
    type?: string | null;
    locations?: Array<string | null>;
    actions?: Array<string | null>;
    identifier?: string | null;
    privileges?: Array<string | null>;
    resources?: Array<string | null>;
};
// ...
export type SapScaiGrantsAuthorizationServiceAuthorizationRequests = {
    ID?: string;
    createdAt?: string | null;
    createdBy?: string | null;
    modifiedAt?: string | null;
    modifiedBy?: string | null;
    client_id?: string | null;
    redirect_uri?: string | null;
    request_uri?: string | null;
    scope?: string | null;
    state?: string | null;
    code_challenge?: string | null;
    code_challenge_method?: string | null;
    grant_management_action?: string | null;
    authorization_details?: string | null;
    requested_actor?: string | null;
    expires_at?: string | null;
    status?: string | null;
    response_type?: string | null;
    risk_level?: string | null;
    subject_token_type?: string | null;
    subject_token?: string | null;
    subject?: string | null;
    expires_in?: number | null;
    grant?: SapScaiGrantsAuthorizationServiceGrants | null;
    grant_id?: string | null;
    consent?: SapScaiGrantsAuthorizationServiceConsents | null;
    access?: Array<SapScaiGrantsAuthorizationServiceAuthorizationRequestsAccess | null>;
};
// ...
export type SapScaiGrantsAuthorizationServiceAuthorizationRequestsAccess = {
    server?: string | null;
    transport?: string | null;
    tools?: SapScaiGrantsAuthorizationServiceCdsMap | null;
    locations?: Array<string | null>;
    actions?: Array<string | null>;
    permissions?: SapScaiGrantsAuthorizationServiceAuthorizationRequestsAccessPermissions | null;
    roots?: Array<string | null>;
    databases?: Array<string | null>;
    schemas?: Array<string | null>;
    tables?: Array<string | null>;
    urls?: Array<string | null>;
    protocols?: Array<string | null>;
    type?: SapScaiGrantsAuthorizationServiceAuthorizationDetailType | null;
    type_code?: string | null;
};
// ...
export type SapScaiGrantsAuthorizationServiceConsents = {
    ID?: string;
    createdAt?: string | null;
    createdBy?: string | null;
    modifiedAt?: string | null;
    modifiedBy?: string | null;
    grant_id?: string;
    grant?: SapScaiGrantsAuthorizationServiceGrants | null;
    request?: SapScaiGrantsAuthorizationServiceAuthorizationRequests | null;
    request_ID?: string | null;
    client_id?: string | null;
    scope?: string | null;
    authorization_details?: Array<SapScaiGrantsAuthorizationServiceAuthorizationDetails>;
    'authorization_details@count'?: Count;
    duration?: number | null;
    subject?: string | null;
    previous_consent?: SapScaiGrantsAuthorizationServiceConsents | null;
    previous_consent_ID?: string | null;
    previous_consent_grant_id?: string | null;
    redirect_uri?: string | null;
    actor?: string | null;
};
// ...
export type SapScaiGrantsAuthorizationServiceGrants = {
    createdAt?: string | null;
    createdBy?: string | null;
    modifiedAt?: string | null;
    modifiedBy?: string | null;
    id?: string;
    client_id?: string | null;
    risk_level?: string | null;
    status?: 'active' | 'revoked';
    revoked_at?: string | null;
    revoked_by?: string | null;
    subject?: string | null;
    actor?: string | null;
    scope?: string | null;
    consents?: Array<SapScaiGrantsAuthorizationServiceConsents>;
    'consents@count'?: Count;
    requests?: Array<SapScaiGrantsAuthorizationServiceAuthorizationRequests>;
    'requests@count'?: Count;
    authorization_details?: Array<SapScaiGrantsAuthorizationServiceAuthorizationDetails>;
    'authorization_details@count'?: Count;
};
// ...
export type SapScaiGrantsAuthorizationServiceCdsMap = {
    [key: string]: unknown;
};
// ...
export type SapScaiGrantsAuthorizationServiceReturnSapScaiGrantsAuthorizationServicePar = {
    request_uri?: string | null;
    expires_in?: number | null;
};
// ...
export type Count = number | string;